refactor: move GenreAndTagChip component to chips directory and update imports
