+++ frontend/src/__tests__/components/GenreAndTagChip.test.tsx
@@ -4,7 +4,7 @@ import userEvent from '@testing-library/user-event'
 import type { ReactElement } from 'react'
 import { I18nextProvider } from 'react-i18next'
 import { MemoryRouter } from 'react-router-dom'
-import GenreAndTagChip from '../../components/GenreAndTagChip'
+import GenreAndTagChip from '../../components/chips/GenreAndTagChip'
 import i18n from '../../i18n'
 import { getTranslatedRecord } from '../../utils/translations'
