+++ frontend/src/components/production/MetaPanel.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from 'react-i18next'
 import { formatDate } from '../../utils/dateUtils'
 import { getHallDisplayName } from '../../utils/hall'
 import { getLocalizedValue } from '../../utils/localization'
-import GenreAndTagChip from '../GenreAndTagChip'
+import GenreAndTagChip from '../chips/GenreAndTagChip'
 
 import type { CSSProperties } from 'react'
 import type { Production } from '../../types/Productions'
