+++ frontend/src/components/searchbar/SearchControlsBar.tsx
@@ -14,7 +14,7 @@ import {
 } from '@mui/material'
 import { FaSortAmountDown, FaSortAmountUp } from 'react-icons/fa'
 import { useTranslation } from 'react-i18next'
-import GenreAndTagChip from '../GenreAndTagChip'
+import GenreAndTagChip from '../chips/GenreAndTagChip'
 import type { SearchChipOption } from '../../types/GenreAndTagChip'
 import SearchBar from './SearchBar'
 import type { SearchSortDirection, SearchSortTarget, SearchViewMode } from './types'
